Initial Data deployment (#6)
This change is the basis for data deployments.

The configuration is like this:

Your deployment spec has an annotation on your pod in the form: store.autocrane.io/location: /data to specify the root directory for your data.
You add more annotations in the form: data.autocrane.io/data1: autocranegit. This drops autocranegit into /data/data1 in your pod (the data datadeploy (sidecar) or datadeployinit (init) containers pull this data).
The value autocranegit is defined in the datarepo environment variables for DataRepo__Sources. It's set to autocranegit:git@https://github.com/microsoft/AutoCrane.git to specify we use the git data fetch plugin to grab the autocrane sources.
The process works like this:

DataRepo is a service that constantly probes all of its sources, creates tar archives of the latest versions, and builds a manifest file with everything it serves.
Your pod has annotations in the form of data.autocrane.io/data1: autocranegit. AutoCrane will discover that annotation, and add another annotation to your pod in the form of request.data.autocrane.io/data1: <how to fetch it>.
Your data deployment sidecar will fetch that data (using "how to fetch it"--which will call the data repo service for the tarball) and write an annotation status.data.autocrane.io/data1 when it's done.
When there is an upgrade (this isn't implemented), AutoCrane will change the request annotation (doing a safe rollout/taking watchdog or readiness errors into account), the data downloader will get new data, and update the /data/data1 symlink to the new folder.

Other future work would probably be something like an admission controller that automatically injects the data deployment sidecars so you only need to specify the data you want in annotations and the rest is taken care for you. Right now the test workload and data repo uses emptyDir, but I don't know if that will be the case in real life.