Add data deployment syncing sidecar feature (#7)
This change creates a sidecar that syncs data much like the init container does, although the difference is that this continually runs in the background rather than stopping when the first change is done. It monitors the current pod for data annotation requests and processes them as necessary. The status is exposed through a watchdog endpoint so that you can set up watchdog prober to flag the pod when things aren't able to sync.